+++ speeduino/src/decoders/jeep_2000.cpp
@@ -143,7 +143,7 @@ int getCrankAngle_Jeep2000(void)
   }
 
   //Estimate the number of degrees travelled since the last tooth}
-  unsigned long const elapsedTime = lastCrankAngleCalc - tempToothLastToothTime;
+  uint32_t const elapsedTime = lastCrankAngleCalc - tempToothLastToothTime;
   crankAngle += timeToAngleDegPerMicroSec(elapsedTime, degreesPerMicro);
 
   if (crankAngle >= 720)
