+++ speeduino/src/decoders/rover_mems.cpp
@@ -311,8 +311,8 @@ int getCrankAngle_RoverMEMS(void)
     crankAngle += 360;
   }
 
-  unsigned long const lastCrankAngleCalc = micros();
-  unsigned long const elapsedTime = lastCrankAngleCalc - tempToothLastToothTime;
+  uint32_t const lastCrankAngleCalc = micros();
+  uint32_t const elapsedTime = lastCrankAngleCalc - tempToothLastToothTime;
   crankAngle += timeToAngleDegPerMicroSec(elapsedTime, degreesPerMicro);
 
   if (crankAngle >= 720)
